moves specific metrics V2 API tests to own test file
